+++ components/index.js
@@ -53,3 +53,5 @@ import './ui/Button';
 
 import './upload/UploadImage';
 import './upload/UploadInner';
+
+export * from './forms/controls/countries';